# perf(docs): improve doc bundle and load experience