--- a/package.json
+++ b/package.json
@@ -58,7 +58,6 @@
     "babel-preset-react": "^6.5.0",
     "babel-preset-stage-1": "^6.5.0",
     "babel-standalone": "^6.18.1",
-    "brace": "^0.9.0",
     "chai": "^3.5.0",
     "chai-enzyme": "^0.6.0",
     "connect-history-api-fallback": "^1.2.0",