--- a/webpack.config.js
+++ b/webpack.config.js
@@ -83,6 +83,9 @@ webpackConfig.plugins = [
       sui: require('semantic-ui-css/package.json').version,
       faker: require('faker/package.json').version,
       lodash: require('lodash/package.json').version,
+      react: require('react/package.json').version,
+      reactDOM: require('react-dom/package.json').version,
+      jsBeautify: require('js-beautify/package.json').version,
     },
   }),
 ]
@@ -170,11 +173,6 @@ if (__TEST__ || argv.localModules) {
     },
   ]
 } else {
-  // we're not using local modules, we're loading deps via CDNs
-  webpackConfig.entry.vendor = _.without(webpackConfig.entry.vendor, [
-    'faker',
-  ])
-
   // these are browser ready modules or aliased to empty
   // do not parse their source for faster builds
   webpackConfig.module.noParse = [
@@ -191,7 +189,11 @@ if (__TEST__ || argv.localModules) {
   webpackConfig.externals = {
     faker: 'faker',
     'anchor-js': 'AnchorJS',
-    Babel: 'babel-standalone',
+    'babel-standalone': 'Babel',
+    react: 'React',
+    'react-dom': 'ReactDOM',
+    'react-dom/server': 'ReactDOMServer',
+    'html-beautify': 'html_beautify',
   }
 }