Fixes #765 

This PR removes improves the doc bundle size and rebuild time.  I've moved most scripts to CDNs so they can be loaded in parallel and cached longer client side.  I've also added an instantly loaded dimmer to entertain during any wait period.

|            |app.js |    vendor.js | modules |   build   |   rebuild    |
|------------|-------|--------------|---------|-----------|--------------|
| Before     |`3 Mb` | `1,110 Kb`   | `2620`  | `~14s`    | `~2s`        |
| After      |`3 Mb` | `2,018 Kb`   | `2455`  | `~10s`    | `~1s`        |